pyproject.toml
[project]
name = "prior-authorization-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=1.22.0",
    "pydantic>=2.11.7",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "rich>=13.7.0",
    "click>=8.1.7",
    "typer>=0.9.0",
]
